+++ RELEASE-NOTES.md
@@ -35,6 +35,8 @@
   schedulers up. A rolling upgrade would result in no leading scheduler for the duration of the
   roll which could be confusing to monitor and debug.
 - Add a new MTTS (Median Time To Starting) metric in addition to MTTA and MTTR.
+- In addition to CPU resources, RAM resources can now be treated as revocable via the scheduler
+  commandline flag `-enable_revocable_ram`.
 
 ### Deprecations and removals: